requirements.txt
werkzeug==2.1.2
PyMySQL==1.0.2
python-dotenv==0.20.0